[Exporter Refactor] Standardizing MatMulToQLinearMatMul (#1266)
* Standardizing MatMulToQLinearMatMul

* Using log_match
